% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{ALS}
\alias{ALS}
\title{The function performing the Alternating Least Square (ALS) Algorithm.}
\usage{
ALS(DATA, n_R, family, opt = opt, max_ite = max_ite, tol = tol)
}
\arguments{
\item{DATA}{A list. The input data. \code{DATA$y} is the dependent variable.
\code{DATA$X} is the 3-D tensor independent variables.
\code{DATA$W} is other independent variables.}

\item{n_R}{A numerical constant. A predefined value determines the rank of
the approximate matrix}

\item{family}{Family of \kbd{generalized linear model}. Provide three options for model.(see more details in
\strong{Details})}

\item{opt}{Optimization options. Provide two options for optimization
stopping criterion. \strong{opt = 1 or 2}. (see more details in
\strong{Details})}

\item{max_ite}{Maximum iteration. The value of maximum iterations for the
algorithm.}

\item{tol}{Tolerance. The value of tolerance with respect to optimization.}
}
\value{
A list. See \code{tensorReg2D}.
}
\description{
The function performing the Alternating Least Square (ALS) Algorithm.
}
