% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{Calculate_IC_Dev}
\alias{Calculate_IC_Dev}
\title{The function computing information criterion values and deviances.}
\usage{
Calculate_IC_Dev(y, w_seq, df, family)
}
\arguments{
\item{y}{A numerical vector. Dependent variable.}

\item{w_seq}{A numerical vector. Fitted values of the tensor regression model.}

\item{df}{integer. Degree of freedom.}

\item{family}{Family of \kbd{generalized linear model}. Provide three options for model.(see more details in
\strong{Details})}
}
\value{
A list.

  \kbd{IC}: The values of Akaike information criterion (AIC) and
  Bayesian information criterion (BIC).

  \kbd{DoF}: The values of the residual degrees of freedom for the null model and
  the residual degrees of freedom.

  \kbd{Dev_res}: The deviance values of residuals.

  \kbd{Dev}: The deviance values for the null model and the working model.
}
\description{
The function computing information criterion values and deviances.
}
