% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{Check_tidy_input}
\alias{Check_tidy_input}
\title{The function confirming the input variables fit the
requirements of the tensorReg2D function.}
\usage{
Check_tidy_input(y, X, W, n_R, family)
}
\arguments{
\item{y}{A numerical vector. Dependent variable.}

\item{X}{A numerical 3-D array. Independent variable(3-D tensor).}

\item{W}{A numerical matrix. Independent variable.}

\item{n_R}{A numerical constant. A predefined value determines the rank of
the approximate matrix}

\item{family}{Family of \kbd{generalized linear model}. Provide three options for model.(see more details in
\strong{Details})}
}
\value{
A list.
  \kbd{DATA}: list of input data.\code{DATA$y} is the dependent variable.
    \code{DATA$X} is the 3-D tensor independent variables.
    \code{DATA$W} is other independent variables.

  \kbd{fm}: The model expression shown in \code{summary.tsglm}.
}
\description{
The function confirming the input variables fit the
requirements of the tensorReg2D function.
}
