% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.tsglm.R
\name{drawpixelmarks}
\alias{drawpixelmarks}
\title{drawpixelmarks}
\usage{
drawpixelmarks(img, marks)
}
\arguments{
\item{img}{an image data.}

\item{marks}{an 0, -1, 1 marking on the image data.}
}
\value{
\kbd{plot.tsglm} returns xxx.
}
\description{
\kbd{drawpixelmarks} .
}
\details{
\kbd{drawpixelmarks} is
}
\examples{
# 

}
\seealso{
\code{\link{TRtest.omics}, \link{plot.tsglm}}
}
\author{
Ping-Yang Chen
}
