% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{getGLMCoef}
\alias{getGLMCoef}
\title{getGLMCoef: Computing the regression coefficients of generalized linear model.}
\usage{
getGLMCoef(X, y, family, offset_vec)
}
\arguments{
\item{X}{A numerical matrix. Independent variable.}

\item{y}{A numerical vector. Dependent variable.}

\item{family}{Family of \kbd{generalized linear model}. Provide three options for model.(see more details in
\strong{Details})}

\item{offset_vec}{A numerical vector. Prior knowledge to be included in the predictors.}
}
\value{
list.
}
\description{
getGLMCoef: Computing the regression coefficients of generalized linear model.
}
