% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{\%b\%}
\alias{\%b\%}
\title{Computation of two matrices: the column of the output matrix is the
Kronecker product of two columns of each input matrix.}
\usage{
A \%b\% B
}
\arguments{
\item{A}{A numerical matrix.}

\item{B}{A numerical matrix.}
}
\value{
numerical matrix.
}
\description{
Computation of two matrices: the column of the output matrix is the
Kronecker product of two columns of each input matrix.
}
