% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{\%hp\%}
\alias{\%hp\%}
\title{Computation of the vector of Hadamard product values of the matrices
\code{X[,,i]} and \code{B}.}
\usage{
X \%hp\% B
}
\arguments{
\item{X}{A numerical 3D array. Each slice is a matrix of size the same as \code{B}.}

\item{B}{A numerical matrix.}
}
\value{
numerical vector.
}
\description{
Computation of the vector of Hadamard product values of the matrices
\code{X[,,i]} and \code{B}.
}
\author{
Sheng-Mao Chang
}
