% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{\%w\%}
\alias{\%w\%}
\title{Computation of the matrix with rows being the linearized matrix products of
\code{X[,,i]} and \code{B}.}
\usage{
X \%w\% B
}
\arguments{
\item{X}{A numerical 3D array. Each slice is a matrix of column size
the same as the row size of \code{B}.}

\item{B}{A numerical matrix.}
}
\value{
numerical 3D array.
}
\description{
Computation of the matrix with rows being the linearized matrix products of
\code{X[,,i]} and \code{B}.
}
