% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorReg2D.R
\name{\%wt\%}
\alias{\%wt\%}
\title{Computation of the matrix with rows being the linearized matrix products of
transposed \code{X[,,i]} and \code{B}.}
\usage{
X \%wt\% B
}
\arguments{
\item{X}{A numerical 3D array. Each slice is a matrix of row size
the same as the row size of \code{B}.}

\item{B}{A numerical matrix.}
}
\value{
numerical vector.
}
\description{
Computation of the matrix with rows being the linearized matrix products of
transposed \code{X[,,i]} and \code{B}.
}
