% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.tsglm.R
\name{predict.tsglm}
\alias{predict.tsglm}
\title{Predict by Second-order Tensor Generalized Regression}
\usage{
\method{predict}{tsglm}(object, newx, type = c("link", "response"), ...)
}
\arguments{
\item{object}{Fitted \kbd{"tsglm"} object.}

\item{newx}{a 3-dimensional array for \code{x} for which the prediction are interested.}

\item{type}{the type of prediction required. The default is \code{type = "link"} that returns 
prediction values on the scale of the linear predictors (eta).
Alternatively, set \code{type = "response"} for returning predictions on the scale of the response variable.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
There are two types of the output of \kbd{predict.tsglm} function.
By setting \code{type = "link"}, it returns the values of the linear predictors; 
and by setting \code{type = "response"}, it returns the the expected values of response variable.
For example, for a binomial model, the predictions are log-odds (probabilities on logit scale) 
if \code{type = "link"}, and \code{type = "response"} gives the predicted probabilities of Y=1.
}
\description{
\kbd{predict} method for self-defined class \kbd{"tsglm"}.
}
\examples{
See example codes in \code{\link{TRtest.omics}, \link{summary.tsglm}}.

}
\seealso{
\code{\link{TRtest.omics}, \link{summary.tsglm}}
}
\author{
Ping-Yang Chen
}
