% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TensorTest2D.R
\docType{package}
\name{TensorTest2D}
\alias{TensorTest2D}
\title{The TensorTest2D Package}
\description{
\kbd{TensorTest2D} is a tool used to fit second-order tensor type data.
}
\details{
This package contains functions for fitting generalized linear models on second-order tensor type data.

It mainly focus on parameter estimation, including parameter coefficients and standard deviation.

Currently, this package includes generalized linear models with \kbd{identity link} (Regression),
\kbd{logit link} (Logistic regression), and \kbd{log link} (Poisson regression).

For a complete list of functions, use \kbd{help(package = TensorTest2D)}.
}
\author{
Sheng-Mao Chang \email{smchang110@gm.ntpu.edu.tw}

Wenbin Lu \email{wlu4@ncsu.edu}

Jung-Ying Tzeng \email{jytzeng@ncsu.edu}

Ping-Yang Chen \email{pychen.ping@gmail.com}

Maintainer: Mark Chen \email{l501l501l@gmail.com}
}
